+++ tests/grid_search_test.cpp
@@ -38,75 +38,41 @@ class SimpleCategoricalFunction
   }
 };
 
+#ifdef ENS_USE_COOT
+TEMPLATE_TEST_CASE("GridSearch_SimpleCategoricalFunction", "[GridSearch]",
+    ENS_ALL_TEST_TYPES, arma::imat, coot::imat)
+#else
 TEMPLATE_TEST_CASE("GridSearch_SimpleCategoricalFunction", "[GridSearch]",
     ENS_ALL_TEST_TYPES, arma::imat)
-{
-  // Create and optimize the categorical function with the GridSearch
-  // optimizer.  We must also create a std::vector<bool> that holds the types
-  // of each dimension, and an arma::Row<size_t> that holds the number of
-  // categories in each dimension.
-  SimpleCategoricalFunction c;
-
-  // We have three categorical dimensions only.
-  std::vector<bool> categoricalDimensions;
-  categoricalDimensions.push_back(true);
-  categoricalDimensions.push_back(true);
-  categoricalDimensions.push_back(true);
-
-  // The first category can take 5 values; the second can take 3; the third can
-  // take 12.
-  arma::Row<size_t> numCategories("5 3 12");
-
-  // The initial point for our optimization will be to set all categories to 0.
-  TestType params("0 0 0");
-
-  // Now create the GridSearch optimizer with default parameters, and run the
-  // optimization.
-  // The GridSearch type can be replaced with any ensmallen optimizer that
-  // is able to handle categorical functions.
-  GridSearch gs;
-  gs.Optimize(c, params, categoricalDimensions, numCategories);
-
-  REQUIRE(params(0) == 0);
-  REQUIRE(params(1) == 2);
-  REQUIRE(params(2) == 1);
-}
-
-
-#ifdef ENS_HAVE_COOT
-
-TEMPLATE_TEST_CASE("GridSearch_SimpleCategoricalFunction", "[GridSearch]",
-    coot::mat, coot::fmat, coot::imat)
-{
-  // Create and optimize the categorical function with the GridSearch
-  // optimizer.  We must also create a std::vector<bool> that holds the types
-  // of each dimension, and an arma::Row<size_t> that holds the number of
-  // categories in each dimension.
-  SimpleCategoricalFunction c;
-
-  // We have three categorical dimensions only.
-  std::vector<bool> categoricalDimensions;
-  categoricalDimensions.push_back(true);
-  categoricalDimensions.push_back(true);
-  categoricalDimensions.push_back(true);
-
-  // The first category can take 5 values; the second can take 3; the third can
-  // take 12.
-  arma::Row<size_t> numCategories("5 3 12");
-
-  // The initial point for our optimization will be to set all categories to 0.
-  TestType params("0 0 0");
-
-  // Now create the GridSearch optimizer with default parameters, and run the
-  // optimization.
-  // The GridSearch type can be replaced with any ensmallen optimizer that
-  // is able to handle categorical functions.
-  GridSearch gs;
-  gs.Optimize(c, params, categoricalDimensions, numCategories);
-
-  REQUIRE(params(0) == 0);
-  REQUIRE(params(1) == 2);
-  REQUIRE(params(2) == 1);
-}
-
 #endif
+{
+  // Create and optimize the categorical function with the GridSearch
+  // optimizer.  We must also create a std::vector<bool> that holds the types
+  // of each dimension, and an arma::Row<size_t> that holds the number of
+  // categories in each dimension.
+  SimpleCategoricalFunction c;
+
+  // We have three categorical dimensions only.
+  std::vector<bool> categoricalDimensions;
+  categoricalDimensions.push_back(true);
+  categoricalDimensions.push_back(true);
+  categoricalDimensions.push_back(true);
+
+  // The first category can take 5 values; the second can take 3; the third can
+  // take 12.
+  arma::Row<size_t> numCategories("5 3 12");
+
+  // The initial point for our optimization will be to set all categories to 0.
+  TestType params("0 0 0");
+
+  // Now create the GridSearch optimizer with default parameters, and run the
+  // optimization.
+  // The GridSearch type can be replaced with any ensmallen optimizer that
+  // is able to handle categorical functions.
+  GridSearch gs;
+  gs.Optimize(c, params, categoricalDimensions, numCategories);
+
+  REQUIRE(params(0) == 0);
+  REQUIRE(params(1) == 2);
+  REQUIRE(params(2) == 1);
+}