+++ tests/iqn_test.cpp
@@ -21,7 +21,7 @@
 using namespace ens;
 using namespace ens::test;
 
-// NOTE: IQN cannot use arma::fp16_mat because pinv() is required.
+// NOTE: IQN cannot use arma::hmat because pinv() is required.
 
 TEMPLATE_TEST_CASE("IQN_LogisticRegressionFunction", "[IQN]", ENS_TEST_TYPES)
 {
@@ -36,15 +36,3 @@ TEMPLATE_TEST_CASE("IQN_LogisticRegressionFunction", "[IQN]", ENS_TEST_TYPES)
         5);
   }
 }
-
-#ifdef ENS_HAVE_COOT
-
-TEMPLATE_TEST_CASE("IQN_LogisticRegressionFunction", "[IQN]",
-    coot::mat, coot::fmat)
-{
-  IQN iqn(0.001, 10, 5000, 0.01);
-  LogisticRegressionFunctionTest<TestType, coot::Row<size_t>>(
-      iqn, 0.003, 0.006);
-}
-
-#endif